pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "amplifier-ipc"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.52",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-timeout>=2.0",
    "pyright>=1.1",
    "ruff>=0.1",
]

[project.scripts]
amplifier-ipc = "amplifier_ipc.cli.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/amplifier_ipc"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "slow: marks tests as slow-running (requires real service subprocesses)",
]

[tool.pyright]
pythonVersion = "3.11"
extraPaths = ["src"]
